+++ src/main/java/com/example/auth_sample/controller/KakaoLoginUIController.java
@@ -1,0 +1,30 @@
+package com.example.auth_sample.controller;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/login")
+public class KakaoLoginUIController {
+
+    @Value("${kakao.client-id}")
+    private String clientId;
+
+    @Value("${kakao.redirect-uri}")
+    private String redirectUri;
+
+    /**
+     * 로그인 페이지
+     */
+    @GetMapping("/page")
+    public String loginPage(Model model) {
+        String location = "https://kauth.kakao.com/oauth/authorize?response_type=code&client_id="
+            + clientId + "&redirect_uri=" + redirectUri;
+
+        model.addAttribute("location", location);
+        return "login";
+    }
+}