+++ CHANGELOG.md
@@ -1,13 +1,41 @@
 ## CHANGE LOG
 
+
+### 4-1) Kakao Social Login 1
+✅ 회원 등록  
+✅ thymeleaf 라이브러리 추가   
+✅ 카카오 소셜 로그인 지원  
+- 카카오에 어플리케이션 등록 &nbsp; `4-1 진행`
+  - kakao developers : https://developers.kakao.com/
+- 간단한 로그인 페이지 구현 &nbsp; `4-1 진행`
+- 코드를 받을 redirect uri path 생성 &nbsp; `4-1 진행`
+- 코드로 카카오에 토큰 요청 &nbsp; `4-2 진행`
+- 받은 토큰으로 사용자 정보 확인  &nbsp; `4-2 진행`
+
+```
+    < 카카오 소셜 로그인 과정 >
+    ㄴ 1. 카카오 로그인 요청 (client id / redirect uri)  
+    ㄴ 2. redirect uri 에 코드를 전달  
+    ㄴ 3. 앱이 코드를 통해 access token을 카카오에 발급 요청  
+    ㄴ 4. access token으로 카카오에 사용자 정보 요청
+```
+
+---
+
 ### 3-1) JPA RelationShip
-✅ jpa w/ relationship  
+✅ jpa w/ relationship
+
+---
 
 ### 2-2) Swagger 연동 & 기본 설정
 ✅ Swagger 연동 & 기본 설정  
 
+---
+
 ### 2-1) JPA 를 활용한 기본 API 구현_1  
 ✅ lombok & simple JPA 구현  
+
+---
 
 ### 1-1) 프로젝트 환경 구성 및 DB 구성
 ✅ Spring Boot 환경 구성